Downloads/program_selecter/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
python-docx
trafilatura
chromadb
sentence-transformers
transformers
torch
openai
